pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppt-template-sdk"
version = "0.1.0"
description = "Reusable PPT template rendering SDK built on python-pptx"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
